requirements.txt
django
south
django-grappelli
pillow
django-filebrowser
django-debug-toolbar
django-menu
django-annoying==0.7.7
